- Renamed files to follow a consistent format - instantworkshop to quick workshop entry, refactoring rgteaching - Fixed sample data fixtures with content type cause by upgrading django (may be an issue when merging with master) (+1 squashed commits)
