+++ myrobogals/rgforums/urls.py
@@ -1,0 +1,34 @@
+from django.conf.urls import url, include, i18n
+from myrobogals.rgforums import views as rgforums_views
+
+# url(r'^forums/newcategory/$', rgforums_views.newcategory),
+# url(r'^forums/newforum/$', rgforums_views.newforum),
+# url(r'^forums/stickytopic/$', rgforums_views.stickytopic),
+# url(r'^forums/setmaxuploadfilesize/$', rgforums_views.setmaxuploadfilesize),
+# url(r'^forums/category/delete/(?P<category_id>\d+)/$', rgforums_views.deletecategory),
+# url(r'^forums/forum/delete/(?P<forum_id>\d+)/$', rgforums_views.deleteforum),
+# url(r'^forums/topic/delete/(?P<topic_id>\d+)/$', rgforums_views.deletetopic),
+# url(r'^forums/post/delete/(?P<post_id>\d+)/$', rgforums_views.deletepost),
+# url(r'^forums/post/fileoffenses/(?P<post_id>\d+)/$', rgforums_views.fileoffenses),
+# url(r'^forums/showforumoffenses/(?P<forum_id>\d+)/$', rgforums_views.showforumoffenses),
+# url(r'^forums/unwatchforum/(?P<forum_id>\d+)/$', rgforums_views.unwatchforum),
+# url(r'^forums/watchforum/(?P<forum_id>\d+)/$', rgforums_views.watchforum),
+# url(r'^forums/unwatchtopic/(?P<topic_id>\d+)/$', rgforums_views.unwatchtopic),
+# url(r'^forums/watchtopic/(?P<topic_id>\d+)/$', rgforums_views.watchtopic),
+# url(r'^forums/unwatchalltopics/$', rgforums_views.unwatchalltopics),
+# url(r'^forums/watchmytopics/$', rgforums_views.watchtopicwithmyposts),
+# url(r'^forums/unwatchall/(?P<uidb36>.+)/(?P<token>.+)/(?P<step>\d)/$', rgforums_views.unwatchall),
+# url(r'^forums/search/(?P<chapterurl>.+)/$', rgforums_views.search),
+# url(r'^forums/newtopic/(?P<forum_id>\d+)/$', rgforums_views.newtopic),
+# url(r'^forums/newpost/(?P<topic_id>\d+)/$', rgforums_views.newpost),
+# url(r'^forums/editpost/(?P<post_id>\d+)/$', rgforums_views.editpost),
+# url(r'^forums/upvote/(?P<post_id>\d+)/$', rgforums_views.topicupvote),
+# url(r'^forums/undoupvote/(?P<post_id>\d+)/$', rgforums_views.topicundoupvote),
+# url(r'^forums/deletefile/(?P<post_id>\d+)/(?P<file_id>\d+)/$', rgforums_views.deletefile),
+# url(r'^forums/downloadpostfile/(?P<post_id>\d+)/(?P<file_id>\d+)/$', rgforums_views.downloadpostfile),
+# url(r'^forums/editforum/(?P<forum_id>\d+)/$', rgforums_views.editforum),
+# url(r'^forums/topic/(?P<topic_id>\d+)/$', rgforums_views.viewtopic),
+# url(r'^forums/forum/(?P<forum_id>\d+)/$', rgforums_views.viewforum),
+# url(r'^forums/$', rgforums_views.forums),
+# url(r'^forums/(?P<forum_id>\d+)/blacklistuser/(?P<user_id>\d+)/$', rgforums_views.blacklistuser),
+# url(r'^forums/(?P<forum_id>\d+)/unblacklistuser/(?P<user_id>\d+)/$', rgforums_views.unblacklistuser'),
