+++ rgtemplates/instant_workshop.html
@@ -1,11 +1,11 @@
 {% extends "base.html" %}
 {% load i18n %}
-{% block title %}{% trans "Create Instant Workshop" %}{% endblock %}
+{% block title %}{% trans "Create Quick Workshop Entry" %}{% endblock %}
 
-{% block pagetitle %}{% trans "Create Instant Workshop" %}{% endblock %}
+{% block pagetitle %}{% trans "Create Quick Workshop Entry" %}{% endblock %}
 
 {% block content %}
-<form action="/teaching/instantworkshop/" method="POST">{% csrf_token %}
+<form action="/teaching/quickentry/" method="POST">{% csrf_token %}
     <table>
   <tr>
     <td colspan="2"><strong>{% trans "Basic information" %}</strong><br />{% trans '<span style="color: #ff0000">*</span> required fields' %}</td>
@@ -70,7 +70,9 @@
     </div>
 
     <p><label for="id_visit_type"><b>Visit type:</b></label> {{ form2.visit_type }}<br /><span style="font-size: 9pt">{% autoescape off %}{{form2.visit_type.help_text}}{% endautoescape %}</span></p>
-
+    {% if form2.errors %}
+	    <div class="error">{% trans "Please correct your entries in the form" %}</div>
+	{% endif %}
     <p> <b>Students taught:</b>
     <table width="460px" style="table-layout:fixed;  text-align:center;" cellpadding="5" border="1" rules="NONE" frame="BOX" cellspacing="0">
 		<col width="100px" />
